pyproject.toml
[project]
name = "mabi-guild-chat-sniffer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "discord-webhook>=1.4.1",
    "pyshark>=0.6",
    "python-dotenv>=1.1.1",
]
